Update to fit new core changes
