--- Assets~/Profiles/TeleportPointerProfile.asset
@@ -1,23 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c156ea3e1a7a493dac236ed84c4a819a, type: 3}
-  m_Name: TeleportPointerProfile
-  m_EditorClassIdentifier: 
-  pointerPrefab: {fileID: 1196247974088106, guid: 0c1e5f0d54ce46319152ece13a4b1eeb,
-    type: 3}
-  pointingExtent: 10
-  pointingRaycastLayerMasks:
-  - serializedVersion: 2
-    m_Bits: 4294967291
-  drawDebugPointingRays: 0
-  debugPointingRayColors:
-  - {r: 0, g: 1, b: 0, a: 1}
