--- Assets~/Profiles/TeleportPointerProfile.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d3736f91cbeeb284fb90b9d389b7fe33
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
