+++ Editor/TeleportInteractorInspector.cs
@@ -2,17 +2,20 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using RealityCollective.Editor.Extensions;
+using RealityToolkit.Editor.UX.Pointers;
 using RealityToolkit.Locomotion.UX;
 using UnityEditor;
 using UnityEngine;
 
 namespace RealityToolkit.Locomotion.Editor
 {
-    [CustomEditor(typeof(ParabolicTeleportPointer))]
-    public class ParabolicTeleportPointerInspector : TeleportPointerInspector
+    [CustomEditor(typeof(TeleportInteractor))]
+    public class TeleportInteractorInspector : LinePointerInspector
     {
         private readonly GUIContent parabolicFoldoutHeaderContent = new GUIContent("Parabolic Pointer Options");
+        private readonly GUIContent teleportFoldoutHeader = new GUIContent("Teleport Pointer Settings");
 
+        private SerializedProperty lineColorAnchor;
         private SerializedProperty minParabolaVelocity;
         private SerializedProperty maxParabolaVelocity;
         private SerializedProperty minDistanceModifier;
@@ -20,7 +23,10 @@ namespace RealityToolkit.Locomotion.Editor
 
         protected override void OnEnable()
         {
+            DrawBasePointerActions = false;
             base.OnEnable();
+
+            lineColorAnchor = serializedObject.FindProperty(nameof(lineColorAnchor));
 
             minParabolaVelocity = serializedObject.FindProperty(nameof(minParabolaVelocity));
             maxParabolaVelocity = serializedObject.FindProperty(nameof(maxParabolaVelocity));
@@ -31,7 +37,15 @@ namespace RealityToolkit.Locomotion.Editor
         public override void OnInspectorGUI()
         {
             base.OnInspectorGUI();
+
             serializedObject.Update();
+
+            if (lineColorAnchor.FoldoutWithBoldLabelPropertyField(teleportFoldoutHeader))
+            {
+                EditorGUI.indentLevel++;
+                EditorGUILayout.PropertyField(lineColorAnchor);
+                EditorGUI.indentLevel--;
+            }
 
             if (minParabolaVelocity.FoldoutWithBoldLabelPropertyField(parabolicFoldoutHeaderContent))
             {
