--- Runtime/UX/TeleportPointer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb5d1a10b6bd4313b861fb62abae6d8f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 08c0fd91bdca45afa09ec64323cf3848, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
